fixup .gitignore to not dig into nsxraml
The current nsx role pulls nsxraml into itself as a sub directory
that is a git managed project. That copy causes the repo tool to
believe things have changed, which pollutes "repo status" calls with
bogus change info. This commit just causes a complete ignore of the
file/ directory, which is where the nrxraml project is placed.

For trees on which the use already ran "repo sync", a complaint may
still arise on "repo status" that .../files/.gitignore has changed.
That is just a cached file relic from the previous invalid .gitignore
in the main project. The developer can just run the following to fix
that:

    git rm --cached files/.gitignore

That will remove tracking on that file due to history.

Change-Id: Ia0f8d10c44d869e8ce65643b195e2743c2415138